feat: 훅 분리 및 구조 개선
